Updates exp, sin, and cos functions for hduals.
Exponential function is highly inefficient. Consider optimizing for
future versions. Added functionality to use exponential with complex
numbers. As of right now, this doesn't work with regular dual numbers.
This is because it is intended only to help calculate sine and cosine.
Consider changing in future versions if there is demand for it. Changed
hdual object so that it can now handle complex numbers.

+++ HDual.py
@@ -118,7 +118,7 @@ class hdual(object):
     
     def __mul__(self, other):
         if type(other) == hdual and other.dim == self.dim:
-            prod = np.zeros(self.dim)
+            prod = np.zeros(self.dim, dtype=self.value.dtype)
             for j in range(self.dim):
                 for k in range(self.dim - j):
                     prod[j + k] += comb(j + k, k)*self.value[j]*other.value[k]
